moved tests of state comparison
