+++ tests/test_jaqsi.py
@@ -43,7 +43,6 @@ from qml_essentials.operations import (
     PhaseDamping,
     ThermalRelaxationError,
 )
-from qml_essentials.math import fidelity, trace_distance, phase_difference
 from qml_essentials.gates import (
     PulseEnvelope,
     PulseInformation,
@@ -1982,271 +1981,6 @@ class TestChunk:
 
         assert chunked_result.shape == full_result.shape
         assert jnp.allclose(full_result, chunked_result, atol=1e-10)
-
-
-class TestFidelity:
-    @pytest.mark.unittest
-    @pytest.mark.parametrize(
-        "sv0,sv1,expected_val",
-        [
-            (
-                jnp.array([1.0, 0.0]),
-                jnp.array([1.0, 0.0]),
-                1.0,
-            ),
-            (
-                jnp.array([1.0, 0.0]),
-                jnp.array([0.0, 1.0]),
-                0.0,
-            ),
-            (
-                jnp.array([0.98753537 - 0.14925137j, 0.00746879 - 0.04941796j]),
-                jnp.array([0.99500417 + 0.0j, 0.09983342 + 0.0j]),
-                None,  # just check against PennyLane
-            ),
-        ],
-        ids=["identical", "orthogonal", "overlap"],
-    )
-    def test_fidelity_statevector(self, sv0, sv1, expected_val):
-        """Fidelity of state vectors should match PennyLane."""
-        result = fidelity(sv0, sv1)
-        expected = qml.math.fidelity_statevector(np.array(sv0), np.array(sv1))
-        assert jnp.allclose(result, expected, atol=1e-10)
-        if expected_val is not None:
-            assert jnp.allclose(result, expected_val, atol=1e-10)
-
-    @pytest.mark.unittest
-    def test_fidelity_statevector_batched(self):
-        """Batched state-vector fidelity should match element-wise PennyLane results."""
-        sv0_batch = jnp.array(
-            [[1.0, 0.0], [0.0, 1.0], [1 / jnp.sqrt(2), 1 / jnp.sqrt(2)]]
-        )
-        sv1_batch = jnp.array(
-            [[1.0, 0.0], [1.0, 0.0], [1 / jnp.sqrt(2), -1 / jnp.sqrt(2)]]
-        )
-        result = fidelity(sv0_batch, sv1_batch)
-        for i in range(3):
-            expected_i = qml.math.fidelity_statevector(
-                np.array(sv0_batch[i]), np.array(sv1_batch[i])
-            )
-            assert jnp.allclose(result[i], expected_i, atol=1e-10)
-
-    @pytest.mark.unittest
-    @pytest.mark.parametrize(
-        "rho0,rho1,expected_val",
-        [
-            (
-                jnp.array([[1, 0], [0, 0]], dtype=jnp.complex128),
-                jnp.array([[1, 0], [0, 0]], dtype=jnp.complex128),
-                1.0,
-            ),
-            (
-                jnp.array([[1, 0], [0, 0]], dtype=jnp.complex128),
-                jnp.array([[0, 0], [0, 1]], dtype=jnp.complex128),
-                0.0,
-            ),
-            (
-                jnp.array([[1, 0], [0, 0]], dtype=jnp.complex128),
-                jnp.eye(2, dtype=jnp.complex128) / 2,
-                None,  # just check against PennyLane
-            ),
-        ],
-        ids=["identical", "orthogonal", "mixed"],
-    )
-    def test_fidelity_dm(self, rho0, rho1, expected_val):
-        """Fidelity of density matrices should match PennyLane."""
-        result = fidelity(rho0, rho1)
-        expected = qml.math.fidelity(np.array(rho0), np.array(rho1))
-        assert jnp.allclose(result, expected, atol=1e-10)
-        if expected_val is not None:
-            assert jnp.allclose(result, expected_val, atol=1e-10)
-
-    @pytest.mark.unittest
-    def test_fidelity_dm_batched(self):
-        """Batched density-matrix fidelity should match element-wise PennyLane."""
-        rho0_batch = jnp.array(
-            [
-                [[1, 0], [0, 0]],
-                [[0.5, 0.5], [0.5, 0.5]],
-                jnp.eye(2) / 2,
-            ],
-            dtype=jnp.complex128,
-        )
-        rho1_batch = jnp.array(
-            [
-                [[0, 0], [0, 1]],
-                [[0.5, 0.5], [0.5, 0.5]],
-                jnp.eye(2) / 2,
-            ],
-            dtype=jnp.complex128,
-        )
-        result = fidelity(rho0_batch, rho1_batch)
-        for i in range(3):
-            expected_i = qml.math.fidelity(
-                np.array(rho0_batch[i]), np.array(rho1_batch[i])
-            )
-            assert jnp.allclose(result[i], expected_i, atol=1e-10)
-
-    @pytest.mark.unittest
-    def test_fidelity_dm_matches_statevector(self):
-        """Density-matrix fidelity of pure states should equal state-vector fidelity."""
-        sv0 = jnp.array([0.98753537 - 0.14925137j, 0.00746879 - 0.04941796j])
-        sv1 = jnp.array([0.99500417 + 0.0j, 0.09983342 + 0.0j])
-        rho0 = jnp.outer(sv0, jnp.conj(sv0))
-        rho1 = jnp.outer(sv1, jnp.conj(sv1))
-        f_sv = fidelity(sv0, sv1)
-        f_dm = fidelity(rho0, rho1)
-        assert jnp.allclose(f_sv, f_dm, atol=1e-8)
-
-    @pytest.mark.unittest
-    def test_fidelity_mismatched_raises(self):
-        """Passing a vector and a matrix should raise ValueError."""
-        sv = jnp.array([1.0, 0.0])
-        dm = jnp.eye(2, dtype=jnp.complex128)
-        with pytest.raises(ValueError, match="same kind"):
-            fidelity(sv, dm)
-
-    @pytest.mark.unittest
-    @pytest.mark.parametrize(
-        "rho0,rho1,expected_val",
-        [
-            (
-                jnp.array([[1, 0], [0, 0]], dtype=jnp.complex128),
-                jnp.array([[1, 0], [0, 0]], dtype=jnp.complex128),
-                0.0,
-            ),
-            (
-                jnp.array([[1, 0], [0, 0]], dtype=jnp.complex128),
-                jnp.array([[0, 0], [0, 1]], dtype=jnp.complex128),
-                1.0,
-            ),
-            (
-                jnp.array([[1, 0], [0, 0]], dtype=jnp.complex128),
-                jnp.eye(2, dtype=jnp.complex128) / 2,
-                None,  # just check against PennyLane
-            ),
-        ],
-        ids=["identical", "orthogonal", "mixed"],
-    )
-    def test_trace_distance(self, rho0, rho1, expected_val):
-        """Trace distance should match PennyLane."""
-        result = trace_distance(rho0, rho1)
-        expected = qml.math.trace_distance(np.array(rho0), np.array(rho1))
-        assert jnp.allclose(result, expected, atol=1e-10)
-        if expected_val is not None:
-            assert jnp.allclose(result, expected_val, atol=1e-10)
-
-    @pytest.mark.unittest
-    def test_trace_distance_batched(self):
-        """Batched trace distance should match element-wise PennyLane."""
-        batch0 = jnp.array(
-            [jnp.eye(2) / 2, jnp.ones((2, 2)) / 2, jnp.array([[1, 0], [0, 0]])],
-            dtype=jnp.complex128,
-        )
-        batch1 = jnp.array(
-            [jnp.ones((2, 2)) / 2, jnp.ones((2, 2)) / 2, jnp.array([[1, 0], [0, 0]])],
-            dtype=jnp.complex128,
-        )
-        result = trace_distance(batch0, batch1)
-        expected = qml.math.trace_distance(np.array(batch0), np.array(batch1))
-        assert jnp.allclose(result, expected, atol=1e-10)
-
-    @pytest.mark.unittest
-    def test_trace_distance_from_statevectors(self):
-        """Trace distance computed from outer-product DMs should match PennyLane."""
-        sv0 = jnp.array([0.2, jnp.sqrt(0.96)])
-        sv1 = jnp.array([1.0, 0.0])
-        rho0 = jnp.outer(sv0, jnp.conj(sv0))
-        rho1 = jnp.outer(sv1, jnp.conj(sv1))
-        result = trace_distance(rho0, rho1)
-        expected = qml.math.trace_distance(np.array(rho0), np.array(rho1))
-        assert jnp.allclose(result, expected, atol=1e-10)
-
-    @pytest.mark.unittest
-    def test_fidelity_2qubit_statevector(self):
-        """Fidelity of 2-qubit state vectors should match PennyLane."""
-        sv0 = jnp.array([1 / jnp.sqrt(2), 0, 0, 1 / jnp.sqrt(2)])  # Bell state
-        sv1 = jnp.array([1, 0, 0, 0], dtype=jnp.complex128)  # |00⟩
-        result = fidelity(sv0, sv1)
-        expected = qml.math.fidelity_statevector(np.array(sv0), np.array(sv1))
-        assert jnp.allclose(result, expected, atol=1e-10)
-
-    @pytest.mark.unittest
-    def test_trace_distance_2qubit(self):
-        """Trace distance of 2-qubit density matrices should match PennyLane."""
-        bell = jnp.array([1 / jnp.sqrt(2), 0, 0, 1 / jnp.sqrt(2)])
-        rho0 = jnp.outer(bell, jnp.conj(bell))
-        rho1 = jnp.eye(4, dtype=jnp.complex128) / 4  # maximally mixed
-        result = trace_distance(rho0, rho1)
-        expected = qml.math.trace_distance(np.array(rho0), np.array(rho1))
-        assert jnp.allclose(result, expected, atol=1e-10)
-
-    @pytest.mark.unittest
-    @pytest.mark.parametrize(
-        "sv0,sv1,expected_val",
-        [
-            (
-                jnp.array([1.0, 0.0]),
-                jnp.array([1.0, 0.0]),
-                0.0,
-            ),
-            (
-                jnp.array([1.0, 0.0]),
-                jnp.array([-1.0, 0.0]),
-                jnp.pi,
-            ),
-            (
-                jnp.array([1.0, 0.0]),
-                jnp.array([1j, 0.0]),
-                jnp.pi / 2,
-            ),
-            (
-                jnp.array([0.98753537 - 0.14925137j, 0.00746879 - 0.04941796j]),
-                jnp.array([0.99500417 + 0.0j, 0.09983342 + 0.0j]),
-                None,  # just check against vdot formula
-            ),
-        ],
-        ids=["identical", "global_phase", "half_pi", "arbitrary"],
-    )
-    def test_phase_difference(self, sv0, sv1, expected_val):
-        """Phase difference should match the vdot formula."""
-        result = phase_difference(sv0, sv1)
-        expected = jnp.angle(jnp.vdot(sv0, sv1))
-        assert jnp.allclose(result, expected, atol=1e-10)
-        if expected_val is not None:
-            assert jnp.allclose(result, expected_val, atol=1e-10)
-
-    @pytest.mark.unittest
-    def test_phase_difference_batched(self):
-        """Batched phase difference should match element-wise computation."""
-        sv0_batch = jnp.array(
-            [
-                [1.0, 0.0],
-                [1.0, 0.0],
-                [1 / jnp.sqrt(2), 1 / jnp.sqrt(2)],
-            ]
-        )
-        sv1_batch = jnp.array(
-            [
-                [1.0, 0.0],
-                [-1.0, 0.0],
-                [1 / jnp.sqrt(2), -1 / jnp.sqrt(2)],
-            ]
-        )
-        result = phase_difference(sv0_batch, sv1_batch)
-        for i in range(3):
-            inner = jnp.sum(jnp.conj(sv0_batch[i]) * sv1_batch[i])
-            expected_i = jnp.angle(inner)
-            assert jnp.allclose(result[i], expected_i, atol=1e-10)
-
-    @pytest.mark.unittest
-    def test_phase_difference_2qubit(self):
-        """Phase difference of 2-qubit states should match the vdot formula."""
-        sv0 = jnp.array([1 / jnp.sqrt(2), 0, 0, 1 / jnp.sqrt(2)])  # Bell |Φ+⟩
-        sv1 = jnp.array([1 / jnp.sqrt(2), 0, 0, 1j / jnp.sqrt(2)])  # phase on |11⟩
-        result = phase_difference(sv0, sv1)
-        expected = jnp.angle(jnp.vdot(sv0, sv1))
-        assert jnp.allclose(result, expected, atol=1e-10)
 
 
 class TestPulse:
